refactor(db): move card entities to shared to fix dependency direction
CardEntity and CardTextEntity model the universal YGOPro .cdb schema (datas/
texts tables), not edopro-specific data, and are consumed cross-context: by the
edopro card repository and by the shared data-source that also backs the mercury
prereleases database. Living under @edopro forced shared/data-source.ts to
import from a bounded context (shared -> context), an inverted dependency. Move
them to src/shared/db/sqlite/infrastructure so shared depends only on shared and
the edopro repository points to shared (context -> shared).
