+++ src/shared/db/sqlite/infrastructure/CardEntity.ts
