+++ src/shared/db/sqlite/infrastructure/CardTextEntity.ts
